Restructure files into `react` folder
