+++ desktop/apps/auction/routes.js
@@ -1,14 +1,14 @@
-import Articles from '../../collections/articles.coffee'
+import Articles from 'desktop/collections/articles.coffee'
 import ArticlesQuery from './queries/articles'
-import Auction from '../../models/auction.coffee'
+import Auction from 'desktop/models/auction.coffee'
 import MeQuery from './queries/me'
 import SaleQuery from './queries/sale'
 import footerItems from './footer_items'
-import metaphysics from '../../../lib/metaphysics'
+import metaphysics from 'lib/metaphysics.coffee'
 import get from 'lodash.get'
-import templateRenderer from '../../lib/template_renderer'
+import templateRenderer from 'desktop/components/react/utils/template_renderer'
 import path from 'path'
-import { getLiveAuctionUrl } from '../../../utils/domain/auctions/urls'
+import { getLiveAuctionUrl } from 'utils/domain/auctions/urls'
 
 export async function index (req, res, next) {
   const { me } = await metaphysics({ query: MeQuery(req.params.id), req: req })
