+++ desktop/components/react/auction_block/utils/test/upcoming_label.js
