+++ desktop/components/react/main_layout/layout.jsx
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import React from 'react'
-import getLayout from '../utils/get_layout'
+import getLayout from './utils/get_layout'
 
 export default function Layout ({ children, ...props }) {
   const { Header, Body, Footer } = getLayout().render(props)
