+++ desktop/components/react/utils/template_renderer.js
